Commit pre HW rework.

+++ Core/Inc/main.h
@@ -55,7 +55,7 @@ void Error_Handler(void);
 
 /* USER CODE BEGIN EFP */
 
-void stdio_parser(uint8_t *ptr, uint32_t len);
+void stdio_parser(uint8_t *ptr, uint16_t len);
 void emergency_stop(void);
 void trigger_json_update(void);
 void dump_packet(uint8_t *data, uint8_t len);
